backend/requirements.txt
# === Dépendances principales ===
fastapi==0.115.2
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3

# === Optionnel (tests et sécurité) ===
pydantic==2.8.2
httpx==0.27.2
